+++ src/elite/libs/EliteSharpLib/EliteSplitAnimatedScreensServiceCollectionExtensions.cs
@@ -1,4 +1,4 @@
-// 'Elite - The Sharp Kind' - Andy Hawkins 2023-2026.
+﻿// 'Elite - The Sharp Kind' - Andy Hawkins 2023-2026.
 // 'Elite - The New Kind' - C.J.Pinder 1999-2001.
 // Elite (C) I.Bell & D.Braben 1984.
 
@@ -89,6 +89,7 @@ internal static class EliteSplitAnimatedScreensServiceCollectionExtensions
             sp.GetRequiredService<GameState>(),
             sp.GetRequiredService<AudioController>(),
             sp.GetRequiredService<IKeyboard>(),
+            sp.GetRequiredService<IGamepad>(),
             sp.GetRequiredService<Stars>(),
             sp.GetRequiredService<PlayerShip>(),
             sp.GetRequiredService<Combat>(),